created debt data model

+++ Dox/Models/Debt+CoreDataProperties.swift
@@ -10,14 +10,16 @@
 import Foundation
 import CoreData
 
+
 extension Debt {
+
     @nonobjc public class func fetchRequest() -> NSFetchRequest<Debt> {
         return NSFetchRequest<Debt>(entityName: "Debt")
     }
 
     @NSManaged public var name: String?
     @NSManaged public var reason: String?
-    @NSManaged public var type: Int
+    @NSManaged public var type: Int16
     @NSManaged public var value: String?
 
 }
